requirements.txt
SpeechRecognition == 3.8.
flask == 1.1.2
urllib3 == 1.25.10
googletrans == 3.0.0
pytesseract == 0.3.6
pillow == 7.2.0
gTTS == 2.1.1
